requirements.txt
accelerate==0.24.1
aiofiles==23.2.1
aiohttp==3.8.6
aiosignal==1.3.1
altair==5.1.2
annotated-types==0.6.0
anyio==3.7.1
arxiv==2.0.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cpm-kernels==1.0.11
cycler==0.12.1
dataclasses-json==0.6.2
deepspeed==0.12.3
distro==1.8.0
exceptiongroup==1.1.3
fastapi==0.104.1
feedparser==6.0.10
ffmpy==0.3.1
filelock==3.13.1
fonttools==4.44.0
frozenlist==1.4.0
fsspec==2023.10.0
gitdb==4.0.11
GitPython==3.1.40
gradio==3.39.0
gradio_client==0.7.0
greenlet==3.0.1
h11==0.14.0
hjson==3.1.0
httpcore==1.0.2
httpx==0.25.1
huggingface-hub==0.19.0
idna==3.4
importlib-metadata==6.8.0
intel-openmp==2021.4.0
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
langchain==0.0.339
langsmith==0.0.66
latex2mathml==3.76.0
linkify-it-py==2.0.2
loguru==0.7.2
Markdown==3.5.1
markdown-it-py==2.2.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.1
mdit-py-plugins==0.3.3
mdtex2html==1.2.0
mdurl==0.1.2
mkl==2021.4.0
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
neo4j-driver==4.1.1
networkx==3.2.1
ninja==1.11.1.1
numpy==1.26.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.52
nvidia-nvtx-cu12==12.1.105
openai==1.3.5
orjson==3.9.10
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
protobuf==4.25.0
psutil==5.9.6
py-cpuinfo==9.0.0
pyarrow==14.0.1
pydantic==2.5.2
pydantic_core==2.14.5
pydeck==0.8.1b0
pydub==0.25.1
Pygments==2.16.1
pynvml==11.5.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-multipart==0.0.6
pytorch-crf==0.7.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
rich==13.6.0
rpds-py==0.12.0
safetensors==0.4.0
semantic-version==2.10.0
sentencepiece==0.1.99
sgmllib3k==1.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.23
sse-starlette==1.6.5
starlette==0.27.0
streamlit==1.28.2
sympy==1.12
tbb==2021.13.0
tenacity==8.2.3
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.1.0+cu118
torchaudio==2.3.1+cu121
torchvision==0.18.1+cu121
tornado==6.3.3
tqdm==4.66.1
transformers==4.30.2
triton==2.1.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
uc-micro-py==1.0.2
urllib3==2.0.7
uvicorn==0.24.0.post1
validators==0.22.0
watchdog==3.0.0
websockets==11.0.3
win32-setctime==1.1.0
yarl==1.9.2
zipp==3.17.0
